backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Kisan AI Backend with Google ADK Integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "fastapi>=0.116.1",
    "firebase-admin>=6.9.0",
    "google-auth>=2.40.3",
    "google-cloud-aiplatform>=1.104.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-speech>=2.33.0",
    "google-cloud-storage>=2.19.0",
    "google-cloud-texttospeech>=2.27.0",
    "google-genai>=1.26.0",
    "google-adk==1.7.0",
    "httpx>=0.28.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn[standard]>=0.35.0",
    "google-cloud-core>=2.4.3",
    "absl-py>=2.2.1",
    "bs4==0.0.2",
    "jinja2==3.1.6",
    "llama-index==0.12.50",
    "requests>=2.32.3",
    "cloudpickle",
]
